requirements.txt
# Core dependencies
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Google Drive API
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Document processing
pypdf==3.17.4
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.4
python-magic==0.4.27
chardet==5.2.0
mammoth==1.6.0  # Better Word document extraction

# Vector storage and embeddings
supabase==2.0.3
openai==1.10.0
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
tiktoken==0.5.2

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Utilities
httpx==0.24.1
tenacity==8.2.3  # For retry logic
schedule==1.2.0  # For scheduling
watchdog==3.0.0  # For file system monitoring

# Logging and monitoring
structlog==24.1.0
rich==13.7.0  # Better console output

# Web interface
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5
python-socketio==5.10.0

# Email alerts
sendgrid==6.11.0  # Optional - for SendGrid
secure-smtplib==0.1.1  # For SMTP email sending

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.1
isort==5.13.2
mypy==1.8.0